fix conditional redundancy for totalBytes
m_progressBar->setValue(qRound(100 * receivedBytes / totalBytes));
There is a potential risk of de-zeroing here, since
receivedBytes / totalBytes will trigger a runtime
error if totalBytes is zero. In addition, if totalBytes
is always greater than or equal to 0, then checking for
totalBytes >= 0 is actually redundant, since the concern
should be that totalBytes is not zero. To ensure the robustness
of the code, a check that totalBytes is zero should be added
to prevent divide-by-zero errors. By ensuring that totalBytes
is greater than zero, this improvement not only avoids the
risk of de-zeroing, but also ensures that the percentage
calculation of the progress bar is meaningful. At the same
time, for cases where totalBytes is 0 or unknown, the progress
bar is displayed as 0% and the format is "unknown size".
This is a safer way to handle the situation and is more
in line with the needs of the actual application.

Pick-to: 6.8 6.7
Change-Id: I7537243e375f336f5c68ef1cc6ec8ade8fa16e5c
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>