Fix static logging category deprecation warnings
This is a second attempt at getting rid of the deprecation warnings
caused by the new logging category macros in 6.9. These cause build
failures on warnings-as-errors configurations, and need to be fixed
properly without affecting backwards compatibility.

The change introduces a new Q_WEBENGINE_LOGGING_CATEGORY macro,
intended to be used across the WebEngine sources. The macro is
not exported to users, so the QtPdf example that declares its
own logging category uses an #ifdef instead.

The new macro is placed in its own header inside src/core/,
and is intended to be removed whenever the next LTS release
allows us to break compatibility with 6.8, and just directly
use Q_STATIC_LOGGING_CATEGORY everywhere.

Fixes: QTBUG-127975
Change-Id: I9174dab21bd597c862e569e170161782cc108ece
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>