CMake: Don't check for the 'xcb' feature in a -no-gui build
When trying to configure qtwebengine with a -no-gui build, the
following error is shown:
CMake Error at qtbase/cmake/QtFeature.cmake:419 (message):
  Attempting to evaluate feature xcb but its definition is missing.
  Either the feature does not exist or a dependency to the module that
  defines it is missing

That happens because the build system never exports
gui/configure.cmake features if Qt was configured with -no-gui, and
yet the webengine configure check tries to reference the non-existent
feature.

Add a if(TARGET Qt6::Gui) condition around the code that checks for
the xcb libs, to avoid trying to evaluate the non-existent feature.

No extra warning is needed, the summary will already mention that Gui
is missing when -no-gui is passed.

Pick-to: 6.8
Fixes: QTBUG-128381
Change-Id: I5f72a8b211653ef6c797eeb1e1440658d625e0e6
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>