Remove assert in RenderWidgetHostViewQt::WheelEventAck
When inspecting with devtools and using the screen cast option,
attempting to scroll in the secondary view will cause the assertion
to fail. This is because the WheelEventAck function gets called twice;
once for the devtools view, and once for the full-size view. This
change removes the assertion, and simply bails out of the function if
no scroll event is pending.

Pick-to: 6.7 6.8
Change-Id: If4518c8be15ca0736756bb1e27a6de6c3d43d198
Reviewed-by: Anu Aliyas <anu.aliyas@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>