Fixup pinch gesture event handling
The current gesture pinch event handling requires the needs_wheel_event
to be set. When the gesture event is created, it is set to true, so that
the input router knows to offer the event to the renderer as a wheel
event. Once the wheel event is acknowledged, needs_wheel_event is set
to false and the event is resent. When the input router receives it the
second time, it knows to send the real gesture event to the renderer.
Qt webengine creates pinch begin and end events with needs_wheel_event
set to true, but the same is missing for update event. The event resent
logic is also missing. This can be done from
RenderWidgetHostViewQt::GestureEventAck() by calling
ForwardTouchpadZoomEventIfNecessary().

Change-Id: I52da137c69da494f497bcd112fcdd466eb08281b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 0dcf9312f9a5cfa87d0392a5c4ddc2d0ebec775e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 2de0ca74cdeb98718a5cafbb76357ced0d107dd1)