Compositor: do not call unbind() from the base destructor
We must call unbind() from the destructor of the final class, not from
the base one. Otherwise, we may encounter various problems when
destroying the compositor or observer due to a race condition since
they are destroyed from different threads.

For example, we might crash on pure virtual function call in Present()
method of NativeSkiaOutputDevice (aka "Compositor") when it's called
from Chrome_InProcGpuThread thread during destruction of a
RenderWidgetHostViewQtDelegateItem (aka "Observer") in the main
thread. There is a call to observer(), which may return a pointer to a
partially destroyed object in such a case.

This is partial revert of 16b2976e2c983cabe6011cd732e3ec46e3402d22.

Fixes: QTBUG-128897
Pick-to: 6.8
Change-Id: Ide0bb8c3d86de7942ae22f2c56382c87199849c7
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
Reviewed-by: Michal Klocek <michal.klocek@qt.io>