Handle missing webrtc in getUserMedia tests
When the webengine_webrtc feature is disabled, desktop media requests
will be delivered as non-desktop media requests instead, which means we
cannot rely on those signals in tests. This shouldn't affect the
previous change to protect against a null desktop media capturer,
because that requires webengine_webrtc=on and
webengine_webrtc_pipewire=off.

Pick-to: 6.8
Change-Id: Id49e41596efd8e0574607a4eb2e0e7b66f5951c9
Reviewed-by: Michael Brüning <michael.bruning@qt.io>