Fix history service reinitalization
setStorageName and setPersistenStoragePath did not
update the storage path of the history service if the
service was already initialized.
It is not supported to change the service's path manually,
so reset the service instead.
The favicon service also needs to be reseted for similar reasons.

This change fixes the flaky test_iconDatabaseMultiView test.

Change-Id: I53fb49832615ea334df30e16167bd32fdaa59038
Reviewed-by: Szabolcs David <davidsz@inf.u-szeged.hu>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>