Doc: Drop unbalanced and unneeded curly brace from code snippet
The QWebEngineUrlSchemeHandler class documentation contains a code
snippet that demonstrates its usage. In a nested conditional, an if
expression opens a scope with a curly brace. This scope is never closed.
Thus, the code is malformed, and copy/pasting the snippet leads to
errors.

As the following statement is a one-liner, there's no need for braces.
Adding braces would be intrusive as Qt's code style mandates that all
branches are braced if any of them are. Therefore, drop the opening
curly brace that follows the if expression.

Fixes: QTBUG-130327
Pick-to: 6.8
Change-Id: I2bb8e8fdde5c554df97e4bcda8b617ef00803de4
Reviewed-by: Jerome Pasion <jerome.pasion@qt.io>