Fix the crash when accessing a blocked site from Google search
When swapping the contents, chromium expects a different SiteInstance
than the current one. When a blocked site is loaded from Google, the new
and current instances were the same, causing the crash. This is because
the error code for the blocked request was set to
net::ERR_BLOCKED_BY_CLIENT. If the request is not browser initiated, it
will use the current process and leading to the crash. Changed the
error code to net::ERR_ACCESS_DENIED to resolve the issue.

Fixes: QTBUG-129796
Change-Id: Id082ddf871357dabcad75d8c8f0d4471c6a77497
Reviewed-by: Kaloyan Chehlarski <kaloyan.chehlarski@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit 474132e4d311377c4c1d94c67c14d3c42a4ed30b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>