Restore default locale at end of tests that set it
Leaking the change to later tests would potentially conceal
locale-dpendencies in them, or cause failures that might appear flaky
due to vanishing on the first rerun (unless the test that set the
locale also happened to be rerun).

Do this using a scope guard, rather than by setting at the end of the
function, since the latter gets missed when the test fails or skips.

Change-Id: Ib756afefaa4e73f47ca78543767da4ffda965286
Reviewed-by: Matthias Rauter <matthias.rauter@qt.io>