Do not warn needlessly about missing dictionaries directory
The warning message will be displayed only when trying to enable spell
checking and the dictionary path is empty. The function has been
modified so that it doesn’t perform any action if the new state and
the current state are the  same.

Fixes: QTBUG-131969
Pick-to: 6.8
Change-Id: Ib69422cdd0285392a6fda967012f9779c950401d
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
(cherry picked from commit a069e0bba44e59c9041205a5432b736bfd980be0)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>