Fix scaling and positioning of popup windows in WebEngineQuick
When using QQuickItem::scale to scale up the UI, popups were misplaced
when a WebEngineView was a child of this scaled item.

After trying different approaches similar to other popups in Qt Quick,
it turned out that the approach that we use to handle rotation can also
be applied to handle scaled items.

This also adds the handling of rotation and scale of items higher up
the item tree.

Pick-to: 6.8
Fixes: QTBUG-130608
Change-Id: If6b6c39113d0f07146d633e9d11d0745f6b0ff15
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit a245afb450d49f4a5f77742fb39f31c44549b134)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>