Fix spurious Wayland+OpenGL assert
The assert at the end of texture() is intended to catch programming
errors which are our responsibility, but also aborts on a minor invalid
operation in qtbase. Log and clear openGL error flags at the start of
the function to limit the assert to guarding only our code.

This also introduces getGLErrorString to log the flag names.

Fixes: QTBUG-132608
Change-Id: I6e7cb9e61757d0290c9cf1bfa5b98a4ffb3e1d52
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
(cherry picked from commit 297e2e20dca52045f6aa65e0311db8d2b9643427)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>