Fix flaky WebEngineViewLoadUrl::test_stopStatus
The test fail was reproducible on arm64 macOS but it could happen on any
platform if it is fast enough.

The value of WebEngineView.url property is just not reliable when a
page's load status is changed to LoadStoppedStatus. If the navigation
events are processed fast enough the navigation request can be committed
right before we call stop() on the view. If the request is
committed it has already updated the url of the current page.

As a fix just ignore the current page url after calling stop in the
test.

Pick-to: 6.8
Change-Id: I9605696dfb1a0e15343afc384fcc2097cfa2994a
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
(cherry picked from commit ccff8d6d518c402759725723e34b5302fdbcf9e1)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>