Fix flaky tst_QWebEnginePage::discardAbortsPendingLoadAndPreservesCommittedLoad
The test fail was reproducible on arm64 macOS but it could happen on any
platform if it is fast enough.

The about:blank page load is interrupted by Discarded lifecycle state in
the test. Loading about:blank can be so fast that the navigation is
committed before we could send the QWebEnginePage::loadStarted signal.
Interrupting the load at this point won't restore the previous commit
and breaks the test.

As a fix, load a page from the disk which load time is potentially
longer than the about:blank page's load time.

Pick-to: 6.8
Change-Id: Id885f8a67b5566c5ad9024e40ce36153ff03f832
Reviewed-by: Anu Aliyas <anu.aliyas@qt.io>
(cherry picked from commit 0730ae7dd64aeef7a23a3a108b95c5057292c56f)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>