Do not use permission objects in tests and fix deprecated apis
Revert tests to use older apis which do not
use permission objects in qml, as those use quint8
enums values which were not yet supported in Qt 6.5
(see QTBUG-112180 for reference)

Also permission policy is not usable as it uses unsupported
enums size. Moreover it seems that we kept ABI with
deprecated apis, but the way it worked changed anyway.
Therefore change the logic back to make compatible with
how it was working in 6.5 -> no persistent permissions.

Change-Id: I0adf5c38075e095259e4e8ff9813e7a3e8357c61
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Kaloyan Chehlarski <kaloyan.chehlarski@qt.io>
Reviewed-by: Michael Brüning <michael.bruning@qt.io>