Turn webEngineCertificateError into a well-behaved value type
Allow a nullptr as the controller and add an "Ok" value to the enum to
signify "no error". This is in line with the "OK" value in Chromium's
net_errors.h. Furthermore, add a public default ctor to allow QMetaType
to create it.

Task-number: QTBUG-108649
Fixes: QTBUG-135032
Pick-to: 6.8
Change-Id: Id06db78f273805ba117eefaa0ada3eac2f538962
Reviewed-by: Michal Klocek <michal.klocek@qt.io>
(cherry picked from commit ba79e85eb5e7e28ff494c95d405164d8b25583d7)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>