Fix accessibility for tables
This change modifies a few of the table-related methods inside
BrowserAccessibilityInterface to use corresponding specialty
methods from Chromium's a11y implementation, instead of relying
on manually extracted attributes. This fixes table contents
always being reported as 'blank' to screen readers.

The kRowHeader and kColumnHeader roles are now exposed as Cell
to Qt (and thus, to the system), which is consistent with what
Chromium does; exposing them to Qt as QAccessible::Row/ColumnHeader
causes some issues when navigating with VoiceOver on macOS.
The caption() method is now implemented as well.

macOS-only implementations of certain platform methods related
to tables have been copied over as well. This fixes some crashes
when iterating through children.

As a drive-by, this also fixes a rare a11y-related crash when in
the middle of deleting a page.

Pick-to: 6.10 6.9 6.8
Fixes: QTBUG-136622
Change-Id: I157bece10d3c2261722d7f70e2b879c1d462b137
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>