Remove logging category macro version checks
The version checks were already wrong, because we also need to support
building 6.11 on top of 6.8, assuming 6.11 is the next LTS. Instead of
fixing them, this commit removes them entirely, to avoid future
submodule update breakage. The explanatory comments have been amended to
correctly reflect when the workaround macros should actually be removed.

Change-Id: Iff894be9697a63eec67ddbe815617b1069e74e3b
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>