tests: fix failing uidelegates tests
On Ubuntu it seems the application is not guaranteed to have focus just
from showing the window. Take a solution from qtdeclarative tests (e.g.
tst_qquicktext, tst_qquickmenu) where we also call requestActivate() and
qWaitForWindowActive() rather than waiting on a focusObject to exist.

Note that this will not change anything on Wayland where window
activation is not supported by default/design. We should probably
skip there; this is also consistent with qtdeclarative tests.

Pick-to: 6.10 6.9
Fixes: QTBUG-137447
Change-Id: Ie19f6dcc0c2b9d1d45805cba8e63e5007f87c115
Reviewed-by: Anu Aliyas <anu.aliyas@qt.io>
Reviewed-by: Michael Brüning <michael.bruning@qt.io>