CMake: Use libtool for archiving QtPdf on static macOS / iOS builds
qtwebengine uses a complicated setup of partial linking and `lipo`
on Apple platforms, to merge Chromium object files and static
libraries into a single multi-arch static library, which is then
linked into the WebEngineCore or Pdf modules.

Partial linking is a process to merge multiple object files into a
single one, which is done by invoking clang -r (which invokes ld -r).

Unfortunately using partial linking on macOS results in losing debug
info. The merged object file contains references to the original
object file debug info, but doesn't contain the debug info itself,
so if the original object files are not available, the debug info is
lost.

To be more precise, if you do the following steps:
- Put the merged object file into a static archive
- Ship the static archive to another machine
- Build a project that links the static archive into an executable
- Run dsymutil on the executable

dsymutil will show a hundreds of warnings like:

  warning: (x86_64) .../pdfium/core/fxcrt/fxcrt/binary_buffer.o
    unable to open object file: No such file or directory

  warning: (x86_64) .../pdfium/core/fxcrt/fxcrt/bytestring.o
    unable to open object file: No such file or directory

Partial linking can also take static libraries as input, not just
object files, but that also results in the same warnings.

For static builds of QtPdf (which we do for our official iOS
packages), we then hit the above described problem.

The issue doesn't happen for shared builds, because the intermediate
static library is merged into the main module shared library on the
build machine, where debug info is still available and dsymutil can
access it.

Instead of using partial linking + lipo to create intermediate
object files or archives, we can use the macOS `libtool` archiver.

It contrast to `ar`, it has support for response files, and it can
also handle a mix of universal or single-arch files as input.

Which means we can feed it a list of GN-generated response files that
contain objects and static libraries, for each arch, and create a
single universal intermediate archive, while also avoiding the need
for `lipo`.

By avoiding partial linking, the debug info is preserved in the final
archive.

To achieve that, we override the internal CMAKE_CXX_ARCHIVE_CREATE
variable to use `libtool` instead of `ar`, and then add the objects
response file directly to the STATIC_LIBRARY_OPTIONS of the static
module library.

One complication with this approach is that `libtool` (in contrast to
`ar`) prints warnings if it finds duplicate object file names during
archiving. These warnings can't be suppressed. They don't seem to
affect the linking process, but they still affect dsymutil which
operates on object names (not paths), and thus will fail to include
some debug info.

To avoid that, we run a cmake script that creates a copy of the
response file, but replaces the object paths with symlinks that
contain a unique name for each of the object files. The archiver
uses then symlink names as the object names, thus avoiding duplicates.

And to ensure rebuilds work correctly, we use a small workaround
using an OBJECT library to create a dependency chain between the
QtPdf static library and the new response file. We employ some of the
old genex tricks to avoid leaking the OBJECT library into the
INTERFACE_LINK_LIBRARIES and .prl files of the QtPdf module.

Unfortunately this `libtool` approach can't be used for macOS shared
builds. We would have to rename the object names used by static
libraries that GN itself creates, and we don't have control over
their naming without modifying GN.
There is a WIP change for that, but it's more invasive and incomplete.

For non-static builds of macOS, we use the previously existing
approaches:
- For universal macOS shared builds, we still create intermediate
  archives using partial linking, and lipo them later.
- For single-arch macOS shared builds, we link the response files
  directly to the module target, like for the other shared library
  platforms.

We don't support shared builds of iOS + qtwebengine, so that's not
relevant.

Fixes: QTBUG-116619
Change-Id: I39c6f542d376f129982f1a323fbdad4a8dff5e56
Reviewed-by: Alexey Edelev <semlanik@gmail.com>
(cherry picked from commit e0aafdd92f708f6b75ad78faecdb7134fc96e966)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>