Add setting to trim accessibility identifiers
By default, Qt's accessibility code will provide an object's
full accessible hierarchy inside the Identifier string that's
passed to a screen reader or automation software. This may
be undesirable for users who use WebEngine for automated
testing.

This change adds a new QWebEngineSetting that disables this
behavior, and instead makes sure that the HTML 'id'
attribute is passed cleanly to accessibility software.
When there is no 'id' attribute, the default Qt behavior
still applies, and the Identifier string is still the
full object hierarchy.

Fixes: QTBUG-134762
Change-Id: Ica0b626c05b5cdca0bad83280ba15183ff9480f6
Reviewed-by: Michael Brüning <michael.bruning@qt.io>