Add userInitiated into QWebEngineNavigationRequest
NavigationType alone is not always sufficient to determine whether
a navigation was triggered by the user. For example,
LinkClickedNavigation can be initiated by JavaScript (window.location
or click() on an anchor), and a RedirectNavigation may carry a user
gesture if it originated from a user-initiated request. The existing
API does not expose this information.

Add a new read-only property `userInitiated` to
QWebEngineNavigationRequest (C++ and QML), similar to the one already
available in QWebEngineNewWindowRequest. The value is derived from
Chromium's HasUserGesture() and IsRendererInitiated() on the
NavigationHandle: a navigation is considered user-initiated when it
has a user gesture or is browser-initiated (address bar, bookmarks,
back/forward).

Note: a redirect navigation that was originally triggered by a user
gesture may also be reported as user-initiated, because Chromium
propagates the gesture flag through the redirect chain.

[ChangeLog][QtWebEngineCore][QWebEngineNavigationRequest] Added
userInitiated property to indicate whether the navigation was
triggered by a user gesture (keyboard or mouse event).
[ChangeLog][QtWebEngineQuick][WebEngineNavigationRequest] Added
userInitiated property to indicate whether the navigation was
triggered by a user gesture (keyboard or mouse event).

Fixes: QTBUG-138126
Change-Id: I49953d2e9df712eeb67523713bb028f7a909d52f
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>