Tests: fortify window activation tests
Some tests require window activation so they can interact and query
objects with active focus. For those, merely exposing the window is not
enough because Ubuntu's default WM settings do not guarantee activation
for new windows. We request that explicitly and wait on activation.

We also need to disable those tests on platforms that don't support
QWindow::requestActivate(), add a new header with a macro based on
qtdeclarative tests to support that.

Pick-to: 6.10 6.9 6.8
Change-Id: I7d80310e2318f9001b2dd5c581209e75b0c599fe
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>