Add permission element resources
HTMLPermissionElement was introduced in the 126 update.
It comes with a set of new localized strings which are
generated at build time.
Add the missing pak file to webengine.

Fixes FATAL:resource_bundle.cc(1210)] Check failed: !data->empty()
crash on google meet when trying to allow mic access in a video call.

Pick-to: 6.10
Change-Id: Id3f7c27b145e63656d79df6004a48b2ed63a7e13
Reviewed-by: Szabolcs David <davidsz@inf.u-szeged.hu>