Add test mock delegate plugin
Previews patch changed the way ui delegates are loaded.
Even though all delegates were tested on device it failed on ci.
The reason was that test mock delegate still relayed
on old 'way' of loading delegates. As we now load plugins,
change test ui delegates to be a plugin.

Add new qml module which contains both TestMockDelegates
and TestParams components and does not depend on Qt Quick Controls.
Missing delegates are taken from WebEngine.ControlsDelegates
as it was before.

Note we do not use plugin in this case and we link with backend
library statically as CI compiles test with qt-cmake-standalone-tests,
but does not install anything, meaning plugin would not be
deployed on testing node. The alternative would require
moving mock delegate 'plugin' to src directory, which looks messy.

Set proper QTWEBENGINE_UI_DELEGATE_MODULE for qml_tests.

Fix import for tst_uidelegates.

Pick-to: 6.10
Change-Id: I01b6d58f10028378933272446c2afc3b98dcfed9
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>