Stabilize scrollPosition auto test on fractional zoom level devices
This test fails for example when the system has a 125% global zoom level
set.

Inside Chromium: The requested pixel positions will be scaled up by the
device pixel ratio (1.25). The resulting fractional number will be
floored down and used for scrolling.

In RenderWidgetHostViewQt: We are informed about the scroll event and
the non-fractional integer is provided for us. We try to scale this
value down by dividing with the DPR to get back the requested pixel
position - but the result will never be accurate becase of the lost
fractional part in Chromium.

In this example, the difference between the requested and the actual
scroll position is  outside of the tolerance level of qFuzzyCompare.
The good news is Chrome behaves the same way, it provides the same
inaccurate fractional scroll position in JavaScript as WebEngine
(both in the API and in JavaScript). So the best we can do is
relaxing the test by implementing a less strict comparison function
for floats.

Pick-to: 6.10 6.9
Change-Id: I2e924fac4344126727c347098b7f17da700181e0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>