Fix race condition in basicFilterOverHTTP
Fav icon service is not bound to overall navigation status
meaning, page can be loaded, but fav service can be
still fetching icons. Moreover if new fav icon request comes
to favicon service and the previous failed request did not manage
to set WasUnableToDownloadFavicon the new request is fired.
This means that tests which are handcrafted and count http
request can suffer from race condition as there is currently
no way to check it fav icon service has completed all the
processing.

To workaround the issue simply add fav.png to basicFilterOverHTTP
test case and wait for iconUrlChanged signal, this way
we can be sure the fav icon http fetches are done.

Note, the change however requires updating of number of
fired calls in spies, as before the change fav icon was considered
not downloadable (unless it was run on windows due to mentioned race
condition), now we have fav icon so any reload will trigger new fetch.

This should not have effect on overall logic if the test.

Pick-to: 6.10
Fixes: QTBUG-138736
Change-Id: Icdd4652becd39ba52dd392901907399817b9ae18
Reviewed-by: Kaloyan Chehlarski <kaloyan.chehlarski@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>