CMake: Improve deployment or resources on Linux
On Linux we use cmake deployment api and even though
mixing different build types of user application
and Qt itself is allowed and complies fine, it did not work
well with deployment of resources which was just checking
active deployment for resource files. Therefore, add
the fallback option. In case of missing resources of given
type assume the 'candidate' option is always right.

Pick-to: 6.10 6.9
Change-Id: Iec610e2d68f29d2b1242703b2210cc96ff3f7354
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>