QWebEngineExtensionManager: override QObject::event()
A released class can, in general, not add new reimplementations of
virtual functions, because code compiled against the old version will
continue to directly call the old furthest-derived reimplementation
(in the case at hand, that would be QObject::event()). This is called
de-virtualization, and every compiler does it when overrides call the
base class' implementation.

As required by BC rules, add an event() reimplementation. This way, if
in the next version we need to e.g. reimplement timerEvent(), we can
do it instead in event().

Found in API-review.

Pick-to: 6.10 6.10.0
Change-Id: I43386e3cb52a1eb6d80475446f013e4dad283227
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>