Disable GBM if Ozone X11 is not available
If certain X11 libraries are missing then the BUILDFLAG(IS_OZONE_X11) is
false. In this case, the GLX code path is unable to create GBM buffer
without Ozone X11 support. This results in the web view being empty and
the warning messages are not helpful.

To circumvent this, add a more meaningful warning message and fallback
to a different rendering path.

Pick-to: 6.10 6.9 6.8
Change-Id: Id4bc9a7e88c6f19713c764437138b92a3aee86e6
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>