Ensure function-scope static variables are only initialized once
Previously some static variables were initialized by checking the value
after the variable was declared, without any guarding mutex. This could
result in multiple threads performing the initialization simultaneously.
In some cases the initialization was idempotent and innocuous, but the
pattern should be avoided, when side effects are introduced there will
be difficult-to-reproduce data races.

Pick-to: 6.10 6.9
Change-Id: I536738dca102cfd1a2a59dd8156084e07ad1922b
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>