Fix user agent override when intercepting a redirect
When a redirect is followed, Chromium will inject the profile's default
user agent string into the list of headers to modify for the redirect
request. This effectively clears any user-provided user agent overrides.

This change removes the User-Agent header from the list of modified
headers, allowing the user-supplied one to be used instead.

Fixes: QTBUG-140515
Pick-to: 6.10 6.9 6.8
Change-Id: I444a286cc5ca728a2d8abcd51eee69cf95a4e543
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>