CMake: Fix ld: library 'GnObject_Pdf_Debug' not found error
When building a single-arch static build of QtPdf on macOS, and
linking to it in a user project with qmake, one got the following
error:

  ld: library 'GnObject_Pdf_Debug' not found

This happened because the build system created an IMPORTED OBJECT
library called GnObject_Pdf_Debug that is linked to the static QtPdf
target via target_link_libraries(PRIVATE).

When linking any dependency privately to a static library, CMake
implicitly also populates its INTERFACE_LINK_LIBRARIES with a
$<LINK_ONLY:GnObject_Pdf_Debug> dependency.

This in turn got processed by qt_generate_prl_file() ->
__qt_internal_walk_libs() which added a plain -lGnObject_Pdf_Debug
to the .prl file of the static QtPdf library, even though we install
no such library.

Make sure we wrap the GnObject_* dependency in a
QT_SKIP_WALK_LIBS_PROCESSING genex, so that __qt_internal_walk_libs
skips adding it to the .prl file.

Wrap the dependency also in a BUILD_INTERFACE genex, so that it's not
exported in the PdfTargets.cmake file either, because the relevant
object file is already included in the static QtPdf library.

Pick-to: 6.8 6.10
Fixes: QTBUG-140032
Change-Id: Ie13e4e473f973564fe7a7f284a646edb18264dc0
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>