Fix illegal iterator dereference in PermissionManagerQt
The converter function between a list of Blink permissions and a list
of Qt permissions, also does some permission merging in appropriate
cases. The logic for this merging was broken, however, because it
assumed the permissions would come in a specific order.

This commit rewrites the whole function, removing the nested for-loops
and potentially unsafe iterator dereferencing.

Fixes: QTBUG-141096
Pick-to: 6.10
Change-Id: If7d207a10a1fe7dd660c21b581ea1ca591fbd429
Reviewed-by: Moss Heim <moss.heim@qt.io>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>