Interrupt printing when QWEPage is destroyed
When a page is destroyed during printing, there is no view to emit
printFinished() and yet the printer thread continues in the background
until the job is finished.

We can improve things somewhat by requesting an interrupt and exiting
early in the printer worker thread. Then at least users do not wait
around forever for the printer to be idle and deleteable.

Add tests both for this case and basic printing with `print(QPrinter*)`

Task-number: QTBUG-140232
Pick-to: 6.8 6.10
Change-Id: If43677b7ad8021d72dd945fd36c3263234692b95
Reviewed-by: Michal Klocek <michal.klocek@qt.io>