NativeSkiaOutputDeviceOpenGL: Fix confusing warning about Vulkan texture
Previously, if dma_buf was not supported or allocating a GBM buffer
object failed, we tried to fall back to a Vulkan texture, assuming the
failure was due to GBM being disabled. However, this fallback would
also fail because a Vulkan texture could not be generated, resulting
in the confusing warning:
"Backend texture is not a Vulkan texture."

This change ensures that the warning is only issued if Chromium actually
generated the texture using Vulkan, avoiding false fallback warnings.

Pick-to: 6.8 6.10
Change-Id: I8866120e6641a1dce846b26f707dc04b489893a7
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>
Reviewed-by: Moss Heim <moss.heim@qt.io>