Examples: Start simplebrowser with OTR profile in single-process mode
Single-process mode doesn’t allow the use of multiple profiles.
Qt WebEngine creates an off-the-record (OTR) profile by default, but
simplebrowser then attempts to create a non-OTR profile for browsing.
Creating a second profile results in a fatal error.

Detect single-process mode, and if it is enabled, make the first
window use the OTR profile. This allows the browser to test
single-process issues without crashing on start.

Change-Id: I1f9d7f151054cc4f1bd27f59924032dc4502b646
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>