QQuickWebEngineView: handle null qmlEngine in navigationRequested()
Bail out earlier and do not crash in navigationRequested callback
when qmlEngine is null. This may happen when using WebEngineView e.g.
in a Loader, as the QML context data is cleared before the object is
destroyed and accordingly, before we unsubscribe from the callback.

Note this is band aid fix, as it is still unclear why this problem
did not occur here before (e.g. with Qt 6.8.3).

Task-number: QTBUG-137768
Change-Id: I33bbaf4240511d20f0b839317fb496fd5b270006
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>
(cherry picked from commit 9359185c706d99e06fd432f0dae21d09e5016321)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>