Make html comoboxes great again (1/2)
During transition from Qt5 to Qt6, we moved some code to
core and in 5d1ef38f9 we moved RWHV delegate to core.
Before that change we used to feed chromium with geometry
and frame geometry of delegated window and later in case
of qquickwindow based delegate we did dome extra 'fixup'
calculations to cover the fact that using geometry()
for view geometry in screen coordinates did not include
actual item delegate offset from the qml scene.

However after the change we use simply mapToGlobal within
the delegate for view geometry which already knows exact
global coordinates for the item within the qml scene.

In e2185036 we have fix for accessibility layout,
however local coordinate mappings makes no sense as it
will always return "0".

this->mapFromItem(this,QPoint(0,0))

Use simply global coordinates here, and remove obsolete
"fixup" calculation for offset in qquickwindow delegate.

Note this patch does not care about rotated qquickwebengineview.
The band-aid patch for that is in followup.

Task-number: QTBUG-135040
Fixes: QTBUG-140321
Fixes: QTBUG-139709
Pick-to: 6.10
Change-Id: I01266b42bc74120e39e2703db78c2b9c3548572f
Reviewed-by: Michael Brüning <michael.bruning@qt.io>
Reviewed-by: Peter Varga <pvarga@inf.u-szeged.hu>