Quick: Add runJavaScript overload with std::function callback
This is to be used in QWebView, where the WebEngine backend is
based on our Quick API. With the addition of a C++ WebView API,
we need a way to call C++ callbacks, and the JavaScript-based callback
mechanism that's currently there is insufficient.

We can't expose an std::function parameter to QML, so this overload
is there for internal use only.

Change-Id: I22b17c2d2ba554078d0e1e39110147399e93102b
Reviewed-by: Michal Klocek <michal.klocek@qt.io>